requirements.txt
argon2-cffi==16.3.0
bcrypt==3.1.4
certifi==2016.2.28
cffi==1.11.2
Django==1.11.3
django-bootstrap3==9.1.0
django-braces==1.11.0
django-debug-toolbar==1.9.1
djangorestframework==3.7.3
Faker==0.8.7
misaka==2.1.0
olefile==0.44
Pillow==4.3.0
pycparser==2.18
python-dateutil==2.6.1
pytz==2017.2
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.1
